+++ CHANGELOG.md
@@ -30,6 +30,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 -   Added Go agent
 -   Added new compiler class `GoCompiler`
 -   Added new stager type `multi_go_exe`
+-   Added go to install script
 
 
 ### Changed